git external image links added to all features

+++ src/features/FeatureAdmin/FeatureAdminHome/FeatureAdminHomeContainer.tsx
@@ -3,8 +3,10 @@ import { createGetStaticProps } from '@/utils/ssgUtils';
 import { FeatureAdminHomeView } from './FeatureAdminHomeView';
 import { DataSliderImages, fetchSliderImages } from '@/features/firebase/api/homeApi';
 import { useContainerData } from '@/hooks/useContainerData';
-import { FirebaseImage, storeImage } from '@/features/firebase/utils/firebaseImageUtils';
+import { deleteExternalImage, ExternalImage, FirebaseImage, storeExternalImage, storeImage } from '@/features/firebase/utils/firebaseImageUtils';
 import { deleteByDbPath } from '@/features/firebase/utils/firebaseGeneralUtils';
+import { uuidV4 } from '@/plugins/uuid';
+import { deleteImageDocument, storeImageDocument } from '@/features/firebase/utils/firebaseDocumentUtils';
 
 const hydrationFncs = [fetchBasicInfo, fetchSliderImages];
 
@@ -21,7 +23,7 @@ export interface FeatureAdminHomeContainerProps extends DataBasicInfo, DataSlide
 export function FeatureAdminHomeContainer({ sliderImages }: FeatureAdminHomeContainerProps) {
   const { state, addToArrayInState, updateState, deleteFromArrayInState, initialLoading } = useContainerData<FeatureAdminMainPageContainerState>(
     { sliderImages: sliderImages ?? [] },
-    hydrationFncs
+    hydrationFncs,
   );
 
   async function handleAddNewImage(image: File) {
@@ -34,13 +36,31 @@ export function FeatureAdminHomeContainer({ sliderImages }: FeatureAdminHomeCont
     }
   }
 
-  async function handleDeleteImage(image: FirebaseImage) {
+  async function handleDeleteImage(image: FirebaseImage | ExternalImage) {
     try {
       updateState({ isDeleteLoading: true });
-      await deleteByDbPath(image.dbPath);
-      deleteFromArrayInState('sliderImages', 'dbPath', image.dbPath);
+      if (image.type === 'firebaseImage') {
+        await deleteByDbPath(image.dbPath);
+        await deleteImageDocument('sliderImages', image.id);
+        deleteFromArrayInState('sliderImages', 'dbPath', image.dbPath);
+      }
+
+      if (image.type === 'externalImage') {
+        await deleteExternalImage({ folder: 'sliderImages', name: image.id });
+        deleteFromArrayInState('sliderImages', 'id', image.id);
+      }
     } finally {
       updateState({ isDeleteLoading: false });
+    }
+  }
+
+  async function handleSaveExternalLink(link: string) {
+    try {
+      updateState({ isSaveLoading: true });
+      const newImage = await storeExternalImage({ folder: 'sliderImages', url: link, name: uuidV4() });
+      addToArrayInState('sliderImages', newImage);
+    } finally {
+      updateState({ isSaveLoading: false });
     }
   }
 
@@ -49,6 +69,7 @@ export function FeatureAdminHomeContainer({ sliderImages }: FeatureAdminHomeCont
       isLoading={initialLoading}
       isUploadLoading={state.isSaveLoading ?? false}
       onUploadImage={handleAddNewImage}
+      onSaveExternalLink={handleSaveExternalLink}
       onRemoveImage={handleDeleteImage}
       images={state.sliderImages}
     />
